0.5.0: support for jj v0.24.0
This just adds `--allow-new` to `jj git push --bookmark <name>` so that
the command will “just work”. Note that this is slightly dangerous if
there are multiple remotes configured. I may revisit this in the future
to account for that myself, including adding a flag to match `jj`, but
likely pending further iteration on jj itself (cf. [discussion][1] on
the change on the jj repo).

[1]: https://github.com/martinvonz/jj/pull/4871#issuecomment-2521250637