0.7.0: Implement JSON Schema with a pattern to control output
This *should* give much more consistent responses, including eliminating
the annoying, all-too-common inclusion of `` ``` `` in the output. In my
testing, this *very* ocasionally fails with weird output from Ollama,
but can always be re-run in that case; but it no longer ever emits
invalid names.

JSON Schema is mildly annoying, but schemars is great!